Move search_files to a tool file (#2098)
